% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_distance.R
\name{calculate_pairwise_distances}
\alias{calculate_pairwise_distances}
\title{Calculate pairwise Euclidean distances}
\usage{
calculate_pairwise_distances(train_coords, test_coords)
}
\arguments{
\item{train_coords}{Matrix of training coordinates (n_train x 2)}

\item{test_coords}{Matrix of test coordinates (n_test x 2)}
}
\value{
Vector of all pairwise distances
}
\description{
Calculate pairwise Euclidean distances
}
\keyword{internal}
