% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.create_coord_matrix}
\alias{.create_coord_matrix}
\title{Create coordinate matrix}
\usage{
.create_coord_matrix(x, y)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}
}
\value{
Matrix with coordinates
}
\description{
Create coordinate matrix
}
\keyword{internal}
