% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_duplicate_coordinates}
\alias{.has_duplicate_coordinates}
\title{Check for duplicate coordinates}
\usage{
.has_duplicate_coordinates(x, y)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}
}
\value{
Logical indicating if duplicates exist
}
\description{
Check for duplicate coordinates
}
\keyword{internal}
