% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_k}
\alias{.validate_k}
\title{Validate k parameter}
\usage{
.validate_k(k, n)
}
\arguments{
\item{k}{Number of folds}

\item{n}{Number of observations}
}
\value{
Validated k value
}
\description{
Validate k parameter
}
\keyword{internal}
