% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spatial_residuals}
\alias{plot_spatial_residuals}
\title{Plot Spatial Residuals}
\usage{
plot_spatial_residuals(residuals_obj, type = "scatter", main = NULL, ...)
}
\arguments{
\item{residuals_obj}{A spatial_residuals object}

\item{type}{Type of plot: "scatter", "histogram", or "qq" (default: "scatter")}

\item{main}{Plot title (default: NULL, auto-generated)}

\item{...}{Additional graphical parameters passed to plot()}
}
\value{
Invisibly returns the residuals object
}
\description{
Creates visualizations of spatial residuals to analyze the spatial
distribution of model errors.
}
\details{
Creates different types of residual plots:
\itemize{
\item "scatter": Spatial scatter plot of residuals colored by magnitude
\item "histogram": Histogram of residual values
\item "qq": Q-Q plot for normality assessment
}
}
\examples{
\dontrun{
observed <- c(1, 2, 3, 4, 5)
predicted <- c(1.1, 2.2, 2.8, 4.1, 4.9)
coords <- cbind(x = c(0, 1, 2, 3, 4), y = c(0, 1, 2, 3, 4))
residuals <- spatial_residuals(observed, predicted, coords)
plot_spatial_residuals(residuals, type = "scatter")
}

}
\seealso{
Other visualization functions:
\code{\link[=plot_spatial_folds]{plot_spatial_folds()}}
}
\concept{visualization functions}
