% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcvR-package.R
\docType{package}
\name{spatialcvR}
\alias{spatialcvR}
\title{spatialcvR: Spatial Cross-Validation for Machine Learning}
\description{
The spatialcvR package provides tools for spatial cross-validation and model
evaluation for geospatial machine learning applications. It addresses spatial
dependence in observations by implementing spatial block, buffered, and
clustering cross-validation methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sowsalim01/spatialcvR}
  \item Report bugs at \url{https://github.com/sowsalim01/spatialcvR/issues}
}

}
\author{
\strong{Maintainer}: Mamadou SOW \email{sowsalim01@gmail.com}

Authors:
\itemize{
  \item Mamadou SOW \email{sowsalim01@gmail.com}
}

}
\keyword{internal}
